chore: restructure file system
