+++ frontend/src/components/common/NavBar/NavMenu.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef, useEffect } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import { LineTexture } from "../../../assets/textures";
-import MenuButton from "./MenuButton";
+import MenuButton from "../Buttons/MenuButton";
 
 export default function NavMenu() {
   const [menuOpen, setMenuOpen] = useState(false);
